pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
gunicorn = "^20.1.0"
django = "^4.2"
envparse = "^0.2.0"
django-extensions = "^3.2.1"
psycopg2-binary = "^2.9.6"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.2"
social-auth-app-django = "^5.2.0"
django-filter = "^23.2"
pydantic = "^1.10.7"
requests = "^2.30.0"
pip = "^23.1.2"
typing-extensions = "^4.5.0"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = 'todolist.settings'
python_files = [
    'tests.py',
    'test_*.py',
    '*_tests.py'
]